+++ src/modules-arch/scheme-registry.js
@@ -4,6 +4,7 @@ let currentScheme = {}
 const SCHEME_IMPORTERS = {
   vanilla: _ => import('@/modules-arch/schemes/vanilla'),
   healthcare: _ => import('@/modules-arch/schemes/healthcare'),
+  reit: _ => import('@/modules-arch/schemes/reit'),
 }
 
 /**