+++ src/vue/modules/dividend-form/store/index.js
@@ -1,0 +1,121 @@
+import { Balance } from '../wrappers/balance'
+import { FEE_TYPES } from '@tokend/js-sdk'
+
+import { types } from './types'
+import { api } from '../_api'
+import { AssetRecord } from '../wrappers/asset.record'
+
+const HORIZON_VERSION_PREFIX = 'v3'
+
+export const state = {
+  accountId: '',
+  balances: [],
+  assets: [],
+  balanceHolders: [],
+}
+
+export const mutations = {
+  [types.SET_ACCOUNT_ID] (state, accountId) {
+    state.accountId = accountId
+  },
+  [types.SET_BALANCES] (state, balances) {
+    state.balances = balances
+  },
+  [types.SET_ASSETS] (state, assets) {
+    state.assets = assets
+  },
+  [types.SET_BALANCE_HOLDERS] (state, balanceHolders) {
+    state.balanceHolders = balanceHolders
+  },
+}
+
+export const actions = {
+  async [types.LOAD_BALANCES] ({ commit, getters }) {
+    const endpoint = `/${HORIZON_VERSION_PREFIX}/accounts/${getters[types.accountId]}`
+    const { data: account } = await api().getWithSignature(endpoint, {
+      include: ['balances.state'],
+    })
+
+    commit(types.SET_BALANCES, account.balances)
+  },
+  async [types.LOAD_ASSETS] ({ commit, getters }) {
+    const endpoint = `/${HORIZON_VERSION_PREFIX}/assets`
+    let { data: assets } = await api().getWithSignature(endpoint, {
+      page: {
+        limit: 100,
+      },
+    })
+
+    commit(
+      types.SET_ASSETS,
+      assets.map(item => new AssetRecord(item, getters[types.balances]))
+    )
+  },
+  /**
+   *
+   * @param {Object} opts
+   * @param {String} opts.assetCode - asset code
+   */
+  async [types.LOAD_BALANCE_HOLDERS] ({ commit }, opts) {
+    const endpoint = `/${HORIZON_VERSION_PREFIX}/balances`
+    const { data: holders } = await api().getWithSignature(endpoint, {
+      filter: {
+        asset: opts.assetCode,
+      },
+      include: ['state'],
+    })
+
+    commit(types.SET_BALANCE_HOLDERS, holders)
+  },
+  /**
+   *
+   * @param {Object} opts
+   * @param {String} opts.assetCode - asset code
+   * @param {String} opts.subtype - subtype fee
+   * @param {String} opts.accountId - account id
+   * @param {String} opts.amount - amount to calculate fee
+   */
+  async [types.LOAD_FEES] ({ commit }, opts) {
+    const endpoint = `/${HORIZON_VERSION_PREFIX}/accounts/${opts.accountId}/calculated_fees`
+    const { data: fees } = await api().getWithSignature(endpoint, {
+      asset: opts.assetCode,
+      fee_type: FEE_TYPES.paymentFee,
+      subtype: opts.subtype,
+      amount: opts.amount || '0',
+    })
+    return fees
+  },
+  /**
+   *
+   * @param {Object} opts
+   * @param {String} opts.accountId - account id
+   */
+  async [types.LOAD_ACCOUNT_ID] ({ commit }, opts) {
+    const endpoint = `/balances/${opts.accountId}/account`
+    const { _rawResponse: account } = await api().getWithSignature(endpoint, {})
+
+    return account.data.account_id
+  },
+}
+
+export const getters = {
+  [types.accountId]: state => state.accountId,
+  [types.balances]: state => state.balances.map(b => new Balance(b)),
+  [types.assets]: state => state.assets.filter(
+    item => item.balance.id && item.isTransferable
+  ),
+  [types.ownedAssets]: state => state.assets.filter(
+    item => item.owner === state.accountId
+  ),
+  [types.balanceHolders]: state => state.balanceHolders
+    .map(item => new Balance(item)),
+}
+
+export const dividendFormModule = {
+  name: 'dividend-form',
+  namespaced: true,
+  state,
+  getters,
+  actions,
+  mutations,
+}