+++ src/vue/pages/assets/AssetDetails.vue
@@ -108,6 +108,25 @@
               </p>
             </td>
           </tr>
+          <tr v-if="asset.maturityDate">
+            <td>
+              {{ 'asset-details.maturity-date' | globalize }}
+            </td>
+            <td>
+              {{ +asset.maturityDate | formatCalendar }}
+            </td>
+          </tr>
+          <tr v-if="asset.annualReturn">
+            <td v-if="ASSET_SUBTYPE.bond === asset.subtype">
+              {{ 'asset-details.annual-return' | globalize }}
+            </td>
+            <td v-else>
+              {{ 'asset-details.expected-revenue' | globalize }}
+            </td>
+            <td>
+              {{ +asset.annualReturn }}%
+            </td>
+          </tr>
         </tbody>
       </table>
     </div>
@@ -153,6 +172,7 @@ import { base } from '@tokend/js-sdk'
 import { ErrorHandler } from '@/js/helpers/error-handler'
 import { Bus } from '@/js/helpers/event-bus'
 
+import { ASSET_SUBTYPE } from '@/js/const/asset-subtypes.const'
 import { mapGetters, mapActions } from 'vuex'
 import { vuexTypes } from '@/vuex'
 
@@ -174,6 +194,7 @@ export default {
     isBalanceCreating: false,
     config,
     EVENTS,
+    ASSET_SUBTYPE,
   }),
   computed: {
     ...mapGetters({